эксп 0 сделан и отлажен, все великолепно

+++ binary.py
@@ -9,7 +9,7 @@ class BinaryUnit:
     def __init__(self, u_radius, sensor_field_radius, etalon, event_diameter, dx,dy):
         self.u_radius = u_radius
         self.sensor_field_radius = sensor_field_radius
-        self.A_min = etalon - event_diameter/2
+        self.A_min = etalon - int(event_diameter/2)
         self.A_max = self.A_min + event_diameter
         self.dx = dx
         self.dy = dy
